chore: remove old scoreboard
