+++ src/pages/home/index.tsx
@@ -4,7 +4,7 @@ import GameContainer from '../../components/containers/GameContainer'
 
 import CoverImg from '../../assets/images/coverImg.png'
 import BouncingArrow from '../../components/common/BouncingArrow'
-import Scoreboard2 from '../../components/scoreboard/Scoreboard2'
+import Scoreboard from '../../components/scoreboard/Scoreboard'
 
 const Home = () => {
   return (
@@ -30,7 +30,7 @@ const Home = () => {
         </div>
         <BouncingArrow />
         <div className='home-options-container'>
-          <Scoreboard2 />
+          <Scoreboard />
           <div className='tutorial-container'>
             <h2>Tips</h2>
           </div>
